src/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asks==3.0.0
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
# attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bidict==0.22.0
biosppy==0.8.0
bleach==5.0.0
# Bottleneck @ file:///tmp/build/80754af9/bottleneck_1648028898966/work
Bottleneck==1.3.7
brotlipy==0.7.0
cachetools==5.0.0
certifi==2022.5.18.1
# cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
# cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
cloudpickle==2.2.1
convertdate==2.4.0
crccheck==1.1
# cryptography @ file:///tmp/build/80754af9/cryptography_1652101588893/work
cryptography==41.0.1
cycler==0.11.0
cytoolz==0.11.0
# dask @ file:///opt/conda/conda-bld/dask-core_1653082102769/work
dask==2023.5.1
databricks-cli==0.17.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docker==5.0.3
docker-pycreds==0.4.0
dos2unix==1
einops==0.4.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.1
Flask==2.1.2
flatbuffers==2.0
fonttools==4.33.3
frozenlist==1.3.0
# fsspec @ file:///opt/conda/conda-bld/fsspec_1652802105984/work
fsspec==2023.5.0
gast==0.5.3
gdown==4.4.0
gitdb==4.0.9
GitPython==3.1.27
gluonts==0.9.4
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.46.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.6.0
hijri-converter==2.2.4
holidays==0.14.2
HydroErr==1.24
hydrostats==0.78
# idna @ file:///tmp/build/80754af9/idna_1637925883363/work
idna==3.4
# imagecodecs @ file:///tmp/build/80754af9/imagecodecs_1635529108216/work
imagecodecs==2023.3.16
imageio==2.19.1
# importlib-metadata==4.11.3 # dask 2023.5.1 depends on importlib-metadata>=4.13.0
importlib-metadata==4.13.0
# iniconfig @ file:///home/linux1/recipes/ci/iniconfig_1610983019677/work
iniconfig==2.0.0
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
ishneholterlib==2020.5.29
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
korean-lunar-calendar==0.2.1
libclang==14.0.1
llvmlite==0.38.1
# locket @ file:///opt/conda/conda-bld/locket_1652903118915/work
locket==1.0.0
lxml==4.9.1
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mega==0.2.11
mistune==0.8.4
mkl-fft==1.3.1
# mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-random==1.2.2
mkl-service==2.4.0
mlflow==1.27.0
multidict==6.0.2
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8
neurokit2==0.2.0
notebook==6.4.11
numba==0.55.2
# numexpr @ file:///tmp/build/80754af9/numexpr_1640689833592/work
numexpr==2.8.4
# numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
numpy
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
outcome==1.2.0
# packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
packaging==23.1
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
# partd @ file:///opt/conda/conda-bld/partd_1647245470509/work
partd==1.4.0
pathlib==1.0.1
pathtools==0.1.2
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==9.1.0
# pluggy @ file:///tmp/build/80754af9/pluggy_1648024445381/work
pluggy==1.0.0
positional-encodings==6.0.1
progressbar2==3.37.1
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.2
promise==2.3
prompt-toolkit==3.0.29
properscoring==0.1
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
# py @ file:///opt/conda/conda-bld/py_1644396412707/work
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
# pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycparser==2.21
pydantic==1.9.1
pyDeprecate==0.3.2
Pygments==2.12.0
PyJWT==2.4.0
PyMeeus==0.5.11
# pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyrsistent==0.18.1
# PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
PySocks==1.7.1
pytest==7.1.1
# pytest-runner @ file:///tmp/build/80754af9/pytest-runner_1623663836848/work
pytest-runner==6.0.0
# python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dateutil==2.8.2
python-utils==2.5.6
pytorch-lightning==1.6.3
pytorchts==0.6.0
pytz==2022.1
# PyWavelets @ file:///tmp/build/80754af9/pywavelets_1648710015787/work
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
querystring-parser==1.2.4
# requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.8
# scikit-image @ file:///tmp/build/80754af9/scikit-image_1648214171611/work
scikit-image
scikit-learn==1.0.2
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
# six @ file:///tmp/build/80754af9/six_1644875935023/work
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.2
stack-data==0.2.0
tabulate==0.8.10
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
terminado==0.15.0
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.0.31.post2005241907
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.1.1
# tomli @ file:///tmp/build/80754af9/tomli_1637314251069/work
tomli==2.0.1
# toolz @ file:///tmp/build/80754af9/toolz_1636545406491/work
toolz==0.12.0
# torch==1.9.0+cu111
torchaudio==0.11.0
torchdiffeq==0.2.3
torchmetrics==0.8.2
# torchvision==0.11.1+cu113
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
trio==0.21.0
typing_extensions==4.2.0
# urllib3 @ file:///opt/conda/conda-bld/urllib3_1650639997961/work
urllib3==2.0.3
wandb==0.12.16
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
wfdb==3.4.1
widgetsnbextension==3.6.0
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
